refactor(eval): move pareto diagnostics out of scripts
Move runtime-probe Pareto analysis into codeminer.eval.agent_runner while keeping scripts/agent_compile/pareto_ci.py as a CLI wrapper.\n\nTesting:\n- pytest test/eval/test_agent_pareto.py test/agent_compile/test_pareto_ci_cli.py test/agent_compile/test_config.py test/agent/test_import_boundaries.py -q\n- pre-commit run --files codeminer/eval/agent_runner/__init__.py codeminer/eval/agent_runner/pareto.py scripts/agent_compile/pareto_ci.py test/eval/test_agent_pareto.py test/agent_compile/test_pareto_ci_cli.py
